At long last, I think this is all working again
In many iterations before this (and since it worked in the video) I was getting an EOF error while building the codespace. I never really figured out what caused it, it was frequently related to installing jupyter kernels, or using commands a user, or from many sequential builds. The best working theory is running out of disk space, likely from cached, sequential builds, or having lots of layers. There are machines with more disk space to try building on Anyway, this one works and if I know what is good for me, I might leave it alone!